requirements.txt
Django>=4.2.0,<6.0.0
Pillow>=10.0.0
django-crispy-forms>=2.0.0
crispy-bootstrap5>=2024.2
python-decouple>=3.8